+++ home.admin/00mainMenu.sh
@@ -171,6 +171,9 @@ fi
 if [ "${labelbase}" == "on" ]; then
   OPTIONS+=(LABELBASE "Labelbase (UTXO labeling)")
 fi
+if [ "${bostr2}" == "on" ]; then
+  OPTIONS+=(BOSTR2 "Bostr2 (Nostr Relay Bouncer)")
+fi
 
 # dont offer to switch to "testnet view for now" - so no wswitch back to mainnet needed
 #if [ ${chain} != "main" ]; then
@@ -333,6 +336,9 @@ case $CHOICE in
             ;;
         LABELBASE)
             sudo /home/admin/config.scripts/bonus.labelbase.sh menu
+            ;;
+        BOSTR2)
+            sudo /home/admin/config.scripts/bonus.bostr2.sh menu
             ;;
         FINTS)
             sudo /home/admin/config.scripts/bonus.fints.sh menu