+++ CHANGES.md
@@ -1,6 +1,7 @@
 ## What's new in Version 1.11.3 of RaspiBlitz?
 
 - New: Tailscale (basic install script) [details](https://en.wikipedia.org/wiki/Tailscale)
+- New: Bostr2 (Nostr Relay Bouncer) [details](https://codeberg.org/Yonle/bostr)
 - Update: LNDK 0.2.0 (Pay BOLT12 offers with LND) [details](https://github.com/lndk-org/lndk/releases/tag/v0.2.0)
 - Update: Helipad (Podcasting 2.0 Boostagram reader) v0.2.0 [details](https://github.com/Podcastindex-org/helipad/releases/tag/v0.2.0)
 - Update: Mempool 3.0.0 [details](https://github.com/mempool/mempool/releases/tag/v3.0.0)